fix(ui): forward size and style in RuntimeObjectIcon helpers
iconForGeneric and iconForSpecialized accepted size / style arguments
but ignored them, so 24pt callers (open-quickly rows) silently got the
default-size icon. Extract the type out of Extensions.swift into its
own file, forward both arguments, and switch the specialised badge from
"G" to "S" so colour-blind users do not have to rely on the teal/pink
distinction alone.
